feat: move video data for the home page to the data folder, fix issues on index pages
