+++ src/pages/index.fr-fr.jsx
@@ -2,13 +2,14 @@
 import React from 'react';
 
 import CaseStudies from 'components/pages/home/case-studies';
-import Community from 'components/pages/home/community';
 import Features from 'components/pages/home/features';
 import Hero from 'components/pages/home/hero';
 import Testimonials from 'components/pages/home/testimonials';
 import WhyEbpf from 'components/pages/home/why-ebpf';
 import Layout from 'components/shared/layout';
 import SEO from 'components/shared/seo';
+import VideoGallery from 'components/shared/video-gallery';
+import videoItems from 'data/home-page/video-data';
 
 const HomePage = () => (
   <Layout>
@@ -17,7 +18,7 @@ const HomePage = () => (
     <WhyEbpf />
     <Testimonials />
     <Features />
-    <Community />
+    <VideoGallery title="eBPF Community Talks" items={videoItems} />
   </Layout>
 );
